+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Spring XML analysis now also recognizes view classes used by [view resolvers](http://docs.spring.io/spring/docs/3.0.x/reference/view.html#view-tiles-url) as _live code_.
 * The custom XML analyzer now allows to specify the predicate `[@attributeName='attributeValue']` in the XPath definition
 * introduced goal `find-only` being the equivalent of `find-without-packaging`, except for having a better name
+* Added analysis of [`faces-config.xml`](http://xmlns.jcp.org/xml/ns/javaee/web-facesconfig_2_2.xsd) files: recognizing listed classes as _live code_.
 
 ## Internal
 * ...